# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vocaseek API
  version: 1.2.0
  description: |
    Dokumentasi API Vocaseek untuk kebutuhan integrasi frontend, QA, dan eksplorasi endpoint lewat Swagger UI.
    Server API memakai path relatif agar tetap cocok di environment lokal maupun deployment.
servers:
  - url: /api
    description: Relative API base URL
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  schemas:
...
    CompanyJobRequest:
      type: object
      properties:
        judul_posisi:
          type: string
        judul_pekerjaan:
          type: string
        deskripsi_pekerjaan:
          type: string
        persyaratan:
          type: string
        lokasi:
          type: string
        tipe_magang:
          type: string
        gaji_per_bulan:
          type: number
        tanggal_penutupan_lamaran:
          type: string
          format: date
        tanggal_mulai_kerja:
          type: string
          format: date
        status:
          type: string
          example: ACTIVE
...
paths:
...
  /intern/profile:
    get:
      tags: [Intern]
      summary: Ambil profil intern
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil intern
...
  /intern/applications:
    get:
      tags: [Intern]
      summary: Daftar lamaran intern
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar lamaran user aktif
  /company/profile:
    get:
      tags: [Company]
      summary: Ambil profil company
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil company
...
  /company/dashboard:
    get:
      tags: [Company]
      summary: Ambil dashboard company
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statistik dashboard
  /company/jobs:
    get:
      tags: [Company]
      summary: Daftar lowongan milik company
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar lowongan
    post:
      tags: [Company]
      summary: Buat lowongan baru
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyJobRequest'
      responses:
        '201':
          description: Lowongan berhasil dibuat
        '403':
          description: Company belum aktif
...